change sentence logic
