+++ packages/frontend/src/utils/utils.ts
@@ -30,3 +30,16 @@ export function renderInlineMarkdown(markdown: string) {
 
   return markdownProcessor.renderInline(markdown)
 }
+
+export function languageJoin(items?: string[]) {
+  if (!items || items.length === 0) {
+    return undefined
+  }
+  if (items.length === 1) {
+    return items[0]
+  }
+  items = [...items]
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const last = items.pop()!
+  return `${items.join(', ')} and ${last}`
+}
